# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CTS OCR API",
    "description": "Cheque Truncation System OCR (Optical Character Recognition) - Mobile and Web API Documentation",
    "version": "1.0.0",
    "contact": {
      "name": "CTS Team",
      "email": "support@cts.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5002",
      "description": "Development Server"
    },
    {
      "url": "http://182.76.112.78:91",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/cts/finalupload": {
      "post": {
        "tags": ["Upload Management"],
        "summary": "Final Upload",
        "description": "Upload finalized batch data",
        "operationId": "finalUpload",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Final upload completed"
          }
        },
        "x-status": "ACTIVE"
      }
    },
...
  }
}